requirements.txt
amqp==2.6.0
asgiref==3.2.3
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
Django==3.0.3
django-cleanup==5.0.0
djangorestframework==3.11.0
future==0.18.2
idna==2.10
kombu==4.6.11
more-itertools==8.4.0
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
pytz==2019.3
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.0
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
